added demicals implementation

+++ Session-06/Calculator/Program.cs
@@ -24,11 +24,11 @@ namespace Calculator
                 case 0: 
                     return num1 + num2;
                 case 1:
-                        return num1 - num2;
+                    return num1 - num2;
                 case 2: 
                     return num1 * num2;
                 case 3:
-                    return num1 / num2;
+                    return (double)num1 / num2;
                 case 4:
                     return Math.Pow(num1,num2);
                 case 5:
